convert dialog_introduction_text to typescript
